refactor(diffusion): fold the eval resolver into the trainer; honest eval-panel claims
Responds to an external (Codex) review of this branch; all findings verified.

- eval_sampling.py had exactly one consumer, so the resolver, retired-key
  table, and cfg_scale_of now live in trainer/diffusion.py — no new module.
- The overlay validated `scheduler` (any object field) as a legal key and
  then died at the first eval ('dict' object has no attribute
  'get_sde_indices'): overlay values are plain YAML, never hydra-instantiated.
  Object fields (scheduler / sde_strategy / sigmas / noise_group_ids /
  init_noise_latent_shape) are now rejected by name at startup.
- The 'same noise, only the policy differs' filmstrip claim holds only at
  eta <= 0: per-step SDE noise is seeded from the eval step's sample ids
  (derive_denoise_step_seed), and both LTX2 recipes run eval_eta 0.7 with
  log_media on. Init now warns for SDE eval and for the disable_driver_xt
  field opt-out; README and _log_eval_media scope the claim to the eta=0
  default. The noise seeding itself is deliberately untouched.
- The EVAL log line printed the eval_samples_per_prompt knob; it now prints
  the resolved value, which eval_sampling.samples_per_prompt can override.

Resolver harness extended to 40 checks (object-field rejections).
